Fix vector_rotate and draw_line.

+++ vector.h
@@ -7,12 +7,12 @@ typedef Point Vector;
 
 Vector vector_summ(Vector, Vector);
 
-Vector vector_multiplication(Vector, double);
+Vector vector_multiplication(Vector, float);
 
 int vector_module(Vector);
 
 Vector vector_normalization(Vector);
 
-Vector vector_rotate(Vector, double);
+Vector vector_rotate(Vector, float);
 
 #endif
